% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_functions.R
\name{taper_volume}
\alias{taper_volume}
\title{Estimate Taper and Volume of a Tree}
\usage{
taper_volume(radii)
}
\arguments{
\item{radii}{A data frame containing tree radii, convex hull areas, circle areas, bark thickness, and other measurements.}
}
\value{
A list with taper information and volume calculations using various methods.
}
\description{
This function estimates the taper and volume of a tree based on RANSAC-fitted radii.
}
