% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_functions.R
\name{veneer_incli}
\alias{veneer_incli}
\title{Correct Tree Inclination Using Stem Axis}
\usage{
veneer_incli(tree_f, plot = TRUE)
}
\arguments{
\item{tree_f}{A LAS object containing tree scan data with `X`, `Y`, and `Z` coordinates.}

\item{plot}{Logical, if `TRUE`, generates and saves a 2D plot comparison of the original and corrected inclination.}
}
\value{
A LAS object with inclination-corrected coordinates (`X_cor`, `Y_cor`, `Z_cor`).
}
\description{
This function estimates and corrects the inclination of a tree point cloud by aligning it to a new coordinate system 
based on the tree's stem axis. It also provides a 2D plot comparison of original and corrected inclination.
}
\examples{
# Assuming 'tree_f' is a valid LAS object:
corrected_tree <- veneer_incli(tree_f, plot = TRUE)
}
