% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_functions.R
\name{veneer_outlier}
\alias{veneer_outlier}
\title{Detect and Remove Outliers from Tree Point Cloud Data}
\usage{
veneer_outlier(radii, tree_f, steps)
}
\arguments{
\item{radii}{A data.frame with calculated radii, X, Y, Z values, and other metrics.}

\item{tree_f}{A LAS object containing tree scan data, with coordinates (X_cor, Y_cor, Z_cor).}

\item{steps}{Numeric, the step size for slicing the stem along the Z-axis.}
}
\value{
A LAS object with outliers removed based on calculated distance deviations.
}
\description{
This function detects outliers based on distance from a fitted spline to the points in a tree's point cloud data.
}
\examples{
# Assuming 'tree_f' is a valid LAS object and 'radii' is computed:
tree_f_cleaned <- veneer_outlier(radii, tree_f, steps = 0.05)
}
